refactor: restructure http-clients folder
